requirements.txt
pandas>=2.3,<3.0
numpy>=2.0.0
PyYAML>=6.0
scipy>=1.3
cvxpy>=1.1.19
ecos>=2.0.14,<3.0.0
empyrical-reloaded>=0.5.0
plotly>=5.0.0,<6.0.0
PyPortfolioOpt>=1.5.0
riskparityportfolio>=0.2
tqdm>=4.65.0
jax>=0.4.0
jaxlib>=0.4.0
